add firebase manager and refactor code
