+++ BookSmart/App/AppDelegate.swift
@@ -1,9 +1,3 @@
-//
-//  AppDelegate.swift
-//  BookSmart
-//
-//  Created by Luka Gazdeliani on 16.01.24.
-//
 
 import UIKit
 import Firebase
