+++ BookSmart/MethodsManager.swift
@@ -1,0 +1,49 @@
+
+import UIKit
+
+final public class MethodsManager {
+    
+    static let shared = MethodsManager()
+    
+    func timeAgoString(from date: Date) -> String {
+        let currentDate = Date()
+        let calendar = Calendar.current
+        let components = calendar.dateComponents([.year, .month, .day, .hour, .minute], from: date, to: currentDate)
+        
+        if let years = components.year, years > 0 {
+            return "\(years)y"
+        } else if let months = components.month, months > 0 {
+            return "\(months)m"
+        } else if let days = components.day, days > 0 {
+            return "\(days)d"
+        } else if let hours = components.hour, hours > 0 {
+            return "\(hours)h"
+        } else if let minutes = components.minute, minutes > 0 {
+            return "\(minutes)m"
+        } else {
+            return "Now"
+        }
+    }
+    
+    func setupLikeCommentShareStackViewUI(_ stackView: UIStackView, imageView: UIImageView, label: UILabel, labelText: String, buttonColor: UIColor, action: Selector) {
+        stackView.axis = .horizontal
+        stackView.alignment = .center
+        stackView.spacing = 8
+        
+        imageView.contentMode = .scaleAspectFit
+        imageView.tintColor = buttonColor
+        imageView.widthAnchor.constraint(equalToConstant: 25).isActive = true
+        imageView.heightAnchor.constraint(equalToConstant: 25).isActive = true
+        
+        label.font = .systemFont(ofSize: 14, weight: .semibold)
+        label.textColor = .customAccentColor
+        label.text = labelText
+        
+        let gestureRecognizer = UITapGestureRecognizer(target: self, action: action)
+        stackView.addGestureRecognizer(gestureRecognizer)
+        
+        stackView.addArrangedSubview(imageView)
+        stackView.addArrangedSubview(label)
+    }
+
+}
