+++ BookSmart/Models/CommentInfo.swift
@@ -1,9 +1,3 @@
-//
-//  CommentInfo.swift
-//  BookSmart
-//
-//  Created by Luka Gazdeliani on 16.01.24.
-//
 
 import Foundation
