+++ BookSmart/Models/PostInfo.swift
@@ -1,9 +1,3 @@
-//
-//  PostInfo.swift
-//  BookSmart
-//
-//  Created by Luka Gazdeliani on 16.01.24.
-//
 
 import Foundation
