+++ BookSmart/Models/UserInfo.swift
@@ -1,9 +1,3 @@
-//
-//  UserInfo.swift
-//  BookSmart
-//
-//  Created by Luka Gazdeliani on 16.01.24.
-//
 
 import Foundation
