+++ BookSmart/Scenes/CreatePost/PostTypeViews/AnnouncementPost/AnnouncementPostView.swift
@@ -1,9 +1,3 @@
-//
-//  AchievementPostView.swift
-//  BookSmart
-//
-//  Created by Luka Gazdeliani on 20.01.24.
-//
 
 import SwiftUI
