+++ BookSmart/Scenes/CreatePost/PostTypeViews/StoryPost/StoryPostView.swift
@@ -1,9 +1,3 @@
-//
-//  StoryPostView.swift
-//  BookSmart
-//
-//  Created by Luka Gazdeliani on 20.01.24.
-//
 
 import SwiftUI
