+++ BookSmart/Scenes/Home/PostsTableViewCell/PostsTableViewCell.swift
@@ -1,9 +1,3 @@
-//
-//  TableViewCell.swift
-//  BookSmart
-//
-//  Created by Luka Gazdeliani on 19.01.24.
-//
 
 import UIKit
 import SwiftUI
@@ -141,7 +135,7 @@ final class PostsTableViewCell: UITableViewCell {
             }
         }
         
-        let timeAgo = viewModel?.timeAgoString(from: postInfo?.postingTime ?? Date())
+        let timeAgo = MethodsManager.shared.timeAgoString(from: postInfo?.postingTime ?? Date())
         
         timeLabel.text = timeAgo
         headerLabel.text = postInfo?.header
@@ -377,7 +371,7 @@ final class PostsTableViewCell: UITableViewCell {
         stackView.addArrangedSubview(label)
     }
     
-    // MARK: - Private Methods
+    // MARK: - Button Methods
     
     @objc private func authorImageTapped(sender: UITapGestureRecognizer) {
 
@@ -429,7 +423,9 @@ final class PostsTableViewCell: UITableViewCell {
     
     @objc private func likeButtonTapped(sender: UITapGestureRecognizer) {
         
-        toggleLikePost()
+        toggleLikePost() { hasLiked in
+            self.updateLikeButtonUI(isLiked: hasLiked)
+        }
         
         if sender.state == .ended {
             UIView.animate(withDuration: 0.1, animations: {
@@ -507,102 +503,36 @@ final class PostsTableViewCell: UITableViewCell {
         }
     }
     
+    // MARK: - Firebase Methods
+    
     func retrieveImage() {
         authorImageView.image = UIImage(systemName: "person.fill")
         authorImageView.tintColor = .customAccentColor
         
-        guard let imageName = postInfo?.authorID.uuidString else { return }
+        guard let postInfoAuthorIDString = postInfo?.authorID.uuidString else { return }
         
-        if let cachedImage = CacheManager.instance.get(name: imageName) {
-            authorImageView.image = cachedImage
-        } else {
-            let database = Firestore.firestore()
-            database.collection("UserInfo").document((postInfo?.authorID.uuidString)!).getDocument { document, error in
-                if error == nil && document != nil {
-                    let imagePath = document?.data()?["image"] as? String
-                    self.fetchImage(imagePath ?? "")
-                }
-            }
-        }
-    }
-    
-    private func fetchImage(_ imagePath: String) {
-        let storageReference = Storage.storage().reference()
-        let fileReference = storageReference.child(imagePath)
-        
-        fileReference.getData(maxSize: 5 * 1024 * 1024) { data, error in
-            if let data = data, error == nil, let fetchedImage = UIImage(data: data) {
-                print("Image fetched successfully.")
-                DispatchQueue.main.async {
-                    self.authorImageView.image = fetchedImage
-                    CacheManager.instance.add(image: fetchedImage, name: self.postInfo?.authorID.uuidString ?? "")
-                }
-            } else {
-                print("Error fetching image:", error?.localizedDescription ?? "Unknown error")
-            }
+        FirebaseManager.shared.retrieveImage(postInfoAuthorIDString) { retrievedImage in
+            self.authorImageView.image = retrievedImage
         }
     }
 
-    private func toggleLikePost() {
+    private func toggleLikePost(completion: @escaping (Bool) -> Void) {
         
         guard let postInfo else { return }
-        
         guard let viewModel else { return }
- 
-        let database = Firestore.firestore()
-        
-        let userReference = database.collection("UserInfo").document(viewModel.userInfo.id.uuidString)
-        let postReference = database.collection("PostInfo").document(postInfo.id.uuidString)
-        
-        let isLiked = viewModel.userInfo.likedPosts.contains(postInfo.id)
-        
-        if isLiked {
-            userReference.updateData([
-                "likedPosts": FieldValue.arrayRemove([postInfo.id.uuidString])
-            ]) { [self] error in
-                if let error = error {
-                    print(error.localizedDescription)
-                    return
-                }
-                
-                postReference.updateData([
-                    "likedBy": FieldValue.arrayRemove([viewModel.userInfo.id.uuidString])
-                ]) { error in
-                    if let error = error {
-                        print(error.localizedDescription)
-                        return
-                    }
-                }
-                updateLikeButtonUI(isLiked: viewModel.userInfo.likedPosts.contains(postInfo.id))
-            }
-        } else {
-            userReference.updateData([
-                "likedPosts": FieldValue.arrayUnion([postInfo.id.uuidString])
-            ]) { [self] error in
-                if let error = error {
-                    print(error.localizedDescription)
-                    return
-                }
-                
-                postReference.updateData([
-                    "likedBy": FieldValue.arrayUnion([viewModel.userInfo.id.uuidString])
-                ]) {  error in
-                    if let error = error {
-                        print(error.localizedDescription)
-                        return
-                    }
-                }
-                updateLikeButtonUI(isLiked: viewModel.userInfo.likedPosts.contains(postInfo.id))
-            }
+          
+        FirebaseManager.shared.toggleLikePost(userInfoIDString: viewModel.userInfo.id.uuidString,
+                                              postInfoIDString: postInfo.id.uuidString) { hasLiked in
+            completion(hasLiked)
         }
     }
 
+    // MARK: - Update UI
+    
     private func updateLikeButtonUI(isLiked: Bool) {
         DispatchQueue.main.async {
             let imageName = isLiked ? "heart.fill" : "heart"
-            
             self.likeButtonImageView.image = UIImage(systemName: imageName)
-            self.likeButtonLabel.textColor = .customLikeButtonColor
         }
     }
 }
