+++ BookSmart/Scenes/TabBarController.swift
@@ -1,9 +1,3 @@
-//
-//  TabBarController.swift
-//  BookSmart
-//
-//  Created by Luka Gazdeliani on 18.01.24.
-//
 
 import UIKit
 import SwiftUI
